Do some renaming and moving files #2710
